pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "toml", "mlcflow"]
build-backend = "setuptools.build_meta"

[project]
name = "mlc-scripts"
description = "Automation scripts for running ML applications using MLC interface"
authors = [
    { name = "MLCommons", email = "systems@mlcommons.org" }
]
license = { file = "LICENSE.md" }
readme = "README.md"
requires-python = ">=3.7"
keywords = ["mlc", "mlcscripts", "mlcflow", "pypi", "package", "automation", "mlperf", "mlcr"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests",
    "pyyaml",
    "mlcflow",
    "giturlparse"
]

[project.urls]
Homepage = "https://github.com/mlcommons/mlperf-automations"
Documentation = "https://docs.mlcommons.org/mlperf-automations"
Repository = "https://github.com/mlcommons/mlperf-automations"
Issues = "https://github.com/mlcommons/mlperf-automations/issues"

[tool.setuptools]
packages = []
include-package-data = true

[tool.setuptools.package-data]
"mlcr" = ["README.md", "VERSION", "git_commit_hash.txt"]
